Moved injector service into sub-folder
